requirements.txt
requests2
pandas
lxml
bs4
selenium
html5lib
python-dotenv
pandera
packaging
pytest
pytest-cov
